Crate bitcoincore_rpc

Rust Client for Bitcoin Core API

This is a client library for the Bitcoin Core JSON-RPC API.

Re-exports

pub extern crate bitcoincore_rpc_json;

pub extern crate jsonrpc;

pub use bitcoincore_rpc_json as json;

pub use json::bitcoin;

Structs

Client
Client implements a JSON-RPC client for the Bitcoin Core daemon or compatible APIs.

Enums

Auth
The different authentication methods for the client.

Error
The error type for errors produced in this library.

Traits

Queryable
A type that can be queried from Bitcoin Core.

RawTx
Used to pass raw txs into the API.

RpcApi

Type Definitions

Result
Crate-specific Result type, shorthand for std::result::Result with our crate-specific Error type;